Ruisth

+++ projetosite/fotos/views.py
@@ -8,14 +8,15 @@ from .models import Categoria, Foto, Utilizador, Comentario
 from django.contrib.auth.decorators import login_required, permission_required, user_passes_test
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User, Group
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404
 from .forms import RegisterForm
 from django.core.exceptions import PermissionDenied
 from django.contrib.auth.decorators import user_passes_test
 
 DEFAULT_PROFILE_URL = '/static/images/profile/default.jpg'
 
-# Create your views here.
+
+##################         FUNCOES        ##############################
 
 # View 403
 def no_perms_page(request):
@@ -32,6 +33,8 @@ def delete_profile_img(utilizador):
     if utilizador.image_url != DEFAULT_PROFILE_URL:
         utilizador.profile_img.delete()
 
+########################################################################
+
 
 # Frontpage
 def index(request):
@@ -42,11 +45,10 @@ def index(request):
         # Aplicar filtro
         fotos = Foto.objects.filter(categoria__nome__contains=categoria)
 
-
     categorias = Categoria.objects.all()
     context = {
         'categorias': categorias,
-        'fotos' : fotos
+        'fotos': fotos
     }
 
     return render(request, 'fotos/index.html', context)
@@ -100,10 +102,17 @@ def galeria(request):
 
 # Lista de utilizadores
 def comunidade(request):
-    utilizadores = Utilizador.objects.all()
+    # Filtro de pesquisa
+    pesquisa = request.GET.get('q')
+
+    if pesquisa is None:
+        utilizadores = Utilizador.objects.all()
+    # Filtro por pesquisa
+    else:
+        utilizadores = Utilizador.objects.filter(user__username__icontains=pesquisa)
 
     context = {
-        'utilizadores': utilizadores
+        'utilizadores': utilizadores,
     }
 
     return render(request, 'fotos/comunidade.html', context)
@@ -113,7 +122,7 @@ def comunidade(request):
 def process_comentario(request, pk):
     comentario = get_object_or_404(Comentario, pk=pk)
     foto_id = comentario.foto.id
-    if comentario.autor == request.user:
+    if comentario.autor == request.user or request.user.is_superuser:
         comentario.delete()
 
     return redirect('fotos:foto', foto_id)
@@ -206,11 +215,11 @@ def criarFoto(request):
             categoria = None
 
         foto = Foto.objects.create(
-            categoria = categoria,
-            titulo = titulo,
-            descricao = descricao,
-            imagem = imagem,
-            autor = request.user
+            categoria=categoria,
+            titulo=titulo,
+            descricao=descricao,
+            imagem=imagem,
+            autor=request.user
         )
 
         return redirect('fotos:galeria')
@@ -293,46 +302,81 @@ def process_logout(request):
     return redirect('fotos:login')
 
 
-@login_required(login_url=reverse_lazy('fotos:login'))
-def profile(request):
+def profile(request, pk):
+    user = get_object_or_404(User, pk=pk)
+    utilizador = get_object_or_404(Utilizador, user_id=user.id)
     likes_count = 0
     fotos = Foto.objects.all()
-    fotos_autor = fotos.filter(autor=request.user.id)
+    fotos_autor = fotos.filter(autor=user.id)
     autor_count = fotos_autor.count()
+    is_following = False
+    if request.user.is_authenticated:
+        utilizador_own = get_object_or_404(Utilizador, user_id=request.user.id)
+    if utilizador.followers.filter(id=request.user.id).exists():
+        is_following = True
+
+    following_list = utilizador.following.all()
+
+    if request.method == 'POST':
+        data = request.POST
+
+        if request.POST.get('follow') is not None and is_following is False:
+            utilizador.followers.add(request.user)
+            utilizador_own.following.add(user)
+        if request.POST.get('unfollow') is not None and is_following is True:
+            utilizador.followers.remove(request.user)
+            utilizador_own.following.remove(user)
+
+        utilizador.save()
+
     for foto in fotos:
-        likes_count += foto.likes.filter(id=request.user.id).count()
+        likes_count += foto.likes.filter(id=user.id).count()
 
     context = {
+        'utilizador': utilizador,
+        'user': user,
         "fotos": fotos_autor,
-        "likes_count": likes_count,
-        "autor_count": autor_count
+        'likes_count': likes_count,
+        'autor_count': autor_count,
+        'is_following': is_following,
+        'following_list': following_list,
     }
+
     return render(request, 'fotos/profile.html', context)
 
 
 @login_required(login_url=reverse_lazy('fotos:login'))
-def profile_edit(request):
+def profile_edit(request, pk):
+    user = get_object_or_404(User, pk=pk)
+    utilizador = get_object_or_404(Utilizador, user_id=pk)
 
     if request.method == 'POST':
         data = request.POST
         user = request.user
-        utilizador = get_object_or_404(Utilizador, user_id=request.user.id)
-        imagem = request.FILES.get('profile_img')
-        primeiro_nome= data['primeiro_nome']
-        ultimo_nome = data['ultimo_nome']
-        about = data['sobre_mim']
+        if request.user == utilizador.user:
+            imagem = request.FILES.get('profile_img')
+            primeiro_nome = data['primeiro_nome']
+            ultimo_nome = data['ultimo_nome']
+            about = data['sobre_mim']
 
-        if primeiro_nome != "":
-            user.first_name = primeiro_nome
-        if ultimo_nome != "":
-            user.last_name = ultimo_nome
-        if about != "":
-            utilizador.about = about
-        if imagem is not None:
-            delete_profile_img(utilizador)
-            utilizador.profile_img = imagem
-        utilizador.save()
-        user.save()
-        return redirect('fotos:profile')
+            if primeiro_nome != "":
+                user.first_name = primeiro_nome
+            if ultimo_nome != "":
+                user.last_name = ultimo_nome
+            if about != "":
+                utilizador.about = about
+            if imagem is not None:
+                delete_profile_img(utilizador)
+                utilizador.profile_img = imagem
+            utilizador.save()
+            user.save()
+            return redirect('fotos:profile', user.id)
 
-    return render(request, 'fotos/profile_edit.html')
+    context = {
+        'utilizador': utilizador,
+        'user': user,
+    }
+
+    return render(request, 'fotos/profile_edit.html', context)
+
+
